go slaytherelics: add some type safety to sync map usage
We currently use sync.Map in a few places, which can be convenient
but has the notable downside of no type information. As a developer
this is pretty hard to interact with, and it can be an easy source of bugs.

Wrapping these in a typed wrapper does not really change anything
about the underlying behavior, but should make it a bit easier for us
humans.