go slaytherelics: add cache to AuthenticateRedis path
This is one of the hottest paths in the service, since each request ends up running a bcrypt operation (which is intentionally very expensive).

This is to compare the supplied token to the hashed version stored in redis. This token is actually an oauth token returned from when we previously authenticated with twitch (and not a user password or secret).

Given this, caching these timebounded token values feels pretty okay and will reduce load on the server by maybe 90% if the profile is to be believed.